requirements.txt
expiringdict==1.1.4
django-model-utils==3.2.0
